+++ pkg/reconciler/broker/resources/names.go
@@ -18,7 +18,7 @@ package resources
 
 import (
 	brokerv1beta1 "github.com/google/knative-gcp/pkg/apis/broker/v1beta1"
-	"github.com/google/knative-gcp/pkg/utils"
+	"github.com/google/knative-gcp/pkg/utils/naming"
 )
 
 // For reference, the minimum number of characters available for a name
@@ -37,26 +37,26 @@ import (
 // Broker. If the topic name would be longer than allowed by PubSub, the
 // Broker name is truncated to fit.
 func GenerateDecouplingTopicName(b *brokerv1beta1.Broker) string {
-	return utils.TruncatedPubsubResourceName("cre-bkr", b.Namespace, b.Name, b.UID)
+	return naming.TruncatedPubsubResourceName("cre-bkr", b.Namespace, b.Name, b.UID)
 }
 
 // GenerateDecouplingSubscriptionName generates a deterministic subscription
 // name for a Broker. If the subscription name would be longer than allowed by
 // PubSub, the Broker name is truncated to fit.
 func GenerateDecouplingSubscriptionName(b *brokerv1beta1.Broker) string {
-	return utils.TruncatedPubsubResourceName("cre-bkr", b.Namespace, b.Name, b.UID)
+	return naming.TruncatedPubsubResourceName("cre-bkr", b.Namespace, b.Name, b.UID)
 }
 
 // GenerateRetryTopicName generates a deterministic topic name for a Trigger.
 // If the topic name would be longer than allowed by PubSub, the Trigger name is
 // truncated to fit.
 func GenerateRetryTopicName(t *brokerv1beta1.Trigger) string {
-	return utils.TruncatedPubsubResourceName("cre-tgr", t.Namespace, t.Name, t.UID)
+	return naming.TruncatedPubsubResourceName("cre-tgr", t.Namespace, t.Name, t.UID)
 }
 
 // GenerateRetrySubscriptionName generates a deterministic subscription name
 // for a Trigger. If the subscription name would be longer than allowed by
 // PubSub, the Trigger name is truncated to fit.
 func GenerateRetrySubscriptionName(t *brokerv1beta1.Trigger) string {
-	return utils.TruncatedPubsubResourceName("cre-tgr", t.Namespace, t.Name, t.UID)
+	return naming.TruncatedPubsubResourceName("cre-tgr", t.Namespace, t.Name, t.UID)
 }
