+++ pkg/reconciler/events/auditlogs/resources/names.go
@@ -21,14 +21,14 @@ import (
 	"fmt"
 
 	"github.com/google/knative-gcp/pkg/apis/events/v1alpha1"
-	"github.com/google/knative-gcp/pkg/utils"
+	"github.com/google/knative-gcp/pkg/utils/naming"
 )
 
 // GenerateTopicName generates a topic name for the audit log
 // source. This refers to the underlying Pub/Sub topic, and not our
 // Topic resource.
 func GenerateTopicName(s *v1alpha1.CloudAuditLogsSource) string {
-	return utils.TruncatedPubsubResourceName("cre-src", s.Namespace, s.Name, s.UID)
+	return naming.TruncatedPubsubResourceName("cre-src", s.Namespace, s.Name, s.UID)
 }
 
 // Generates the resource name for the topic used by an CloudAuditLogsSource.
@@ -39,5 +39,5 @@ func GenerateTopicResourceName(s *v1alpha1.CloudAuditLogsSource) string {
 // GenerateSinkName generates a Stackdriver sink resource name for an
 // CloudAuditLogsSource.
 func GenerateSinkName(s *v1alpha1.CloudAuditLogsSource) string {
-	return utils.TruncatedPubsubResourceName("cre-src", s.Namespace, s.Name, s.UID)
+	return naming.TruncatedPubsubResourceName("cre-src", s.Namespace, s.Name, s.UID)
 }
